SyntaxReader
Note: we don't use a singleton to represent end of list

When dealing with ordinary lists it is acceptable and even desired
to represent all occurances of '() with one instance of the same
object. However when parsing we want to preserve source position
accurately in all cases, so it is important to record the location of
end of lists as well, which can only happen if we use a separate
object instance for every "end-of-list".

So instead of a static object 'SyntaxPair::nil' we add a member function
Syntax::isNil() and a more optimized one SyntaxPair::isNil(). 'nil is
encoded as car=cdr=0.