smalls/codegen/SimpleCodeGen: first implementation, compiles fact.scm
A very hack, primitive implementation walking the AST and generating
truly horrendous code. A few library functions ("+","-","*", etc) are
provided in a hacky way to get it to run.

The generated code conventions will be changed in future commits, but
I will preserve the simpel code generator as a concept. The aim is
to have it generate more "reasonable" code by applying trivial optimi-
zations which can be performed by walking the AST. It should remain
simple and useful as a baseline.

We might even use it to generate IR from the AST.