smalls/ast/Frame: define variable attributes assigned,initialised,etc
Add several attributes to ast::Variable:
- exported
- assignable
- assigned
- initialized
- used

The first three are required to enforce that exported symbols cannot
be assigned to.

All of them can be useful for some primitive optimizations. Obviously
for advanced optimizations used and assigned need to be re-calculated
after transformations, but even these simple attributes are sufficient
for a lot of constant propagation and inlining, etc.